+++ src/edu/clemson/resolve/jetbrains/actions/RunRESOLVEListener.java
@@ -13,11 +13,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Listens for information coming back from the resolve compiler. Captures
- * warnings, errors, and general informational printouts emitted by the
- * compiler. This class, like {@link RunRESOLVEOnLanguageFile} follows very
- * closely the tool listener written for this purpose in the ANTLRv4 plugin
- * written by Terence Parr and others, found here:
+ * Listens for information coming back from the resolve compiler. Captures warnings, errors, and general informational
+ * printouts emitted by the compiler. This class, like {@link RunRESOLVEOnLanguageFile} follows very closely the tool
+ * listener written for this purpose in the ANTLRv4 plugin written by Terence Parr and others, found here:
  * <p>
  * <a href="https://github.com/antlr/intellij-plugin-v4">https://github.com/antlr/intellij-plugin-v4/a>
  */
